Remove iOS Safari audio fix to simplify code
The code specific for fixing audio issues in iOS Safari was removed. This functionality was originally added to create an audio buffer with a desired sample rate specifically for iOS devices, as they have a different default sample rate. However, it was concluded that this fix is unnecessary and it was decided to simplify the code. The change should not impact other functionalities, as it aims at improving code readability and maintenance.